requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
Django==4.1.1
django-async-orm==0.1.10
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
fake-useragent==0.1.11
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
kombu==5.2.4
lxml==4.9.1
Markdown==3.4.1
outcome==1.2.0
parse==1.19.0
prompt-toolkit==3.0.31
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
requests-html==0.10.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
vine==5.0.0
w3lib==2.0.1
wcwidth==0.2.5
websockets==10.3
wsproto==1.2.0
zipp==3.8.1
